test single cell widths
